Amend pets table to add owner

+++ app/Http/Controllers/Api/KangarooController.php
@@ -7,17 +7,27 @@ use App\Http\Requests\CreateKangaroo;
 use App\Http\Requests\UpdateKangaroo;
 use App\Http\Resources\PetResource;
 use App\Models\Pet;
+use App\Models\User;
 use App\QueryBuilders\Kangaroo\SearchFilter;
-use Illuminate\Http\Request;
+use Illuminate\Contracts\Auth\Authenticatable;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 use NadLambino\Uploadable\Actions\Upload;
 use Spatie\QueryBuilder\AllowedFilter;
 use Spatie\QueryBuilder\QueryBuilder;
 
 class KangarooController extends Controller
 {
-    public function index()
+    private User|Authenticatable $user;
+
+    public function __construct()
     {
-        $kangaroos = QueryBuilder::for(Pet::query()->kangaroos())
+        $this->user = auth()->user();
+    }
+
+    public function index(): JsonResponse
+    {
+        $kangaroos = QueryBuilder::for($this->user->pet()->kangaroos())
             ->allowedFilters([
                 AllowedFilter::custom('search', new SearchFilter),
             ])
@@ -28,15 +38,15 @@ class KangarooController extends Controller
         return $this->success(PetResource::collection($kangaroos));
     }
 
-    public function store(CreateKangaroo $request)
+    public function store(CreateKangaroo $request): JsonResponse
     {
         Upload::onlyFor(Pet::class);
-        $kangaroo = Pet::create($request->validated());
+        $kangaroo = $this->user->pet()->create($request->validated());
 
         return $this->success($kangaroo);
     }
 
-    public function update(UpdateKangaroo $request, Pet $kangaroo)
+    public function update(UpdateKangaroo $request, Pet $kangaroo): JsonResponse
     {
         Upload::onlyFor($kangaroo);
         $kangaroo->update($request->validated());
@@ -48,8 +58,10 @@ class KangarooController extends Controller
         return $this->success($kangaroo);
     }
 
-    public function destroy(Pet $kangaroo)
+    public function destroy(Pet $kangaroo): JsonResponse
     {
+        Gate::authorize('delete', $kangaroo);
+
         $kangaroo->delete();
 
         return $this->success([]);
